requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-bdd==7.0.1
pytest-html==4.1.1
pytest-xvfb==3.0.0
pytest-playwright==0.4.3
python-dotenv==1.0.0
